+++ example/frontend/lib/database/sqlite_migrations/migration_v0001.dart
@@ -1,10 +1,10 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // Migration version: 1
-// Generated on 2025-06-04 14:36:40.978685
+// Generated on 2025-06-07 14:22:12.835812
 
 const List<String> migrationSqlStatementsV1 = [
   '''-- Initial Schema (Version 1)''',
-  '''-- Generated on 2025-06-04 14:36:40.973393''',
+  '''-- Generated on 2025-06-07 14:22:12.830431''',
   r'''
 CREATE TABLE IF NOT EXISTS "bookstores" (
   "address" TEXT,