+++ packages/tether/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 1.0.0
 
 - Initial release of the Tether package.
+
+## 1.0.5
+
+- Changed the User Preferences generator to support number arrays.