+++ packages/tether_libs/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 1.0.0
 
 - Initial release of the Tether Libraries.
+
+## 1.0.5
+
+- Changed how the query and realtime are accessed in the CLient Manager.